chore: Extract shared functions to parent
Now that all of the queries share a `conversionGoal` parameter, Python is happy with this!
